backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-docx
groq
pydantic